move masked request log formatters to correct package
